py/tkinter/cupcake-editor-demo/requirements.txt
cupcake-editor==0.25.8
filetype==1.2.0
pillow==11.2.1
Pygments==2.19.2
toml==0.10.2
